# Conflicts:
#	README.md
#	build.fabric-m.gradle.kts
#	build.fabric-o.gradle.kts
#	build.neoforge.gradle.kts
#	settings.gradle.kts
#	src/main/java/lomka/Lomka.java
#	src/main/java/lomka/starl/mixins/com/mojang/blaze3d/platform/MixinNativeImage.java
#	src/main/java/lomka/starl/mixins/com/mojang/blaze3d/platform/MixinTextureUtil.java
#	src/main/java/lomka/starl/mixins/com/mojang/blaze3d/vertex/MixinMeshData.java
#	src/main/java/lomka/starl/mixins/com/mojang/math/MixinQuadrant.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/MixinCamera.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/MixinMinecraft.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/animation/MixinKeyframeAnimation.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/model/MixinModel.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/renderer/MixinDynamicUniformStorage.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/renderer/block/model/MixinFaceBakery.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/renderer/block/model/MixinItemTransform.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/renderer/texture/MixinMipmapGenerator.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/renderer/texture/MixinSpriteContents.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/renderer/texture/MixinTextureAtlas.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/renderer/texture/atlas/sources/MixinPalettedPermutations.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/resources/model/MixinMaterial.java
#	src/main/java/lomka/starl/mixins/net/minecraft/client/sounds/MixinJOrbisAudioStream.java
#	src/main/java/lomka/starl/mixins/net/minecraft/resources/MixinResource.java
#	src/main/java/lomka/starl/mixins/net/minecraft/util/MixinARGB.java
#	src/main/java/lomka/starl/mixins/net/minecraft/util/MixinArrayListDeque.java
#	src/main/java/lomka/starl/mixins/net/minecraft/util/MixinMth.java
#	src/main/java/lomka/starl/mixins/net/minecraft/util/thread/MixinBlockableEventLoop.java
#	src/main/resources/lomka.mixins.json
#	stonecutter.gradle.kts
#	stonecutter.properties.toml
#	versions/1.21-fabric/src/main/resources/1.21.accesswidener
#	versions/1.21-fabric/src/main/resources/lomka.mixins.json
#	versions/1.21-neoforge/src/main/resources/META-INF/accesstransformer.cfg
#	versions/1.21-neoforge/src/main/resources/lomka.mixins.json
#	versions/1.21.11-fabric/src/main/resources/1.21.11.accesswidener
#	versions/1.21.11-neoforge/src/main/resources/META-INF/accesstransformer.cfg
#	versions/1.21.9-fabric/src/main/resources/1.21.9.accesswidener
#	versions/1.21.9-fabric/src/main/resources/lomka.mixins.json
#	versions/1.21.9-neoforge/src/main/resources/META-INF/accesstransformer.cfg
#	versions/1.21.9-neoforge/src/main/resources/lomka.mixins.json
#	versions/26.1-fabric/src/main/resources/26.1.accesswidener
#	versions/26.1-fabric/src/main/resources/lomka.mixins.json
#	versions/26.1-neoforge/src/main/resources/META-INF/accesstransformer.cfg
#	versions/26.1-neoforge/src/main/resources/lomka.mixins.json
#	versions/26.2-fabric/src/main/resources/26.2.accesswidener
#	versions/26.2-fabric/src/main/resources/lomka.mixins.json
#	versions/26.2-neoforge/src/main/resources/META-INF/accesstransformer.cfg
#	versions/26.2-neoforge/src/main/resources/lomka.mixins.json

+++ src/main/java/lomka/starl/mixins/net/minecraft/util/MixinARGB.java
@@ -121,9 +121,9 @@ public class MixinARGB {
         }
 
         int reciprocal = lomka$RECIPROCALS[outA];
-        int r = (rSum * reciprocal) >>> 24;
-        int g = (gSum * reciprocal) >>> 24;
-        int b = (bSum * reciprocal) >>> 24;
+        int r = (int) (((long) rSum * reciprocal) >>> 24);
+        int g = (int) (((long) gSum * reciprocal) >>> 24);
+        int b = (int) (((long) bSum * reciprocal) >>> 24);
 
         return (outA << 24) | (r << 16) | (g << 8) | b;
     }
